cmd/snap-update-ns: move apply{Profile,{User,System}Fstab} to same file
This sets the stage for the upcoming unification of all the three
functions into one.
