+++ cmd/snap-update-ns/export_test.go
@@ -51,17 +51,19 @@ var (
 	// system
 	DesiredSystemProfilePath = desiredSystemProfilePath
 	CurrentSystemProfilePath = currentSystemProfilePath
-	ApplySystemFstab         = applySystemFstab
 
 	// user
 	DesiredUserProfilePath = desiredUserProfilePath
 	CurrentUserProfilePath = currentUserProfilePath
-	ApplyUserFstab         = applyUserFstab
 
 	// xdg
 	XdgRuntimeDir        = xdgRuntimeDir
 	ExpandPrefixVariable = expandPrefixVariable
 	ExpandXdgRuntimeDir  = expandXdgRuntimeDir
+
+	// update
+	ApplySystemFstab = applySystemFstab
+	ApplyUserFstab   = applyUserFstab
 )
 
 // SystemCalls encapsulates various system interactions performed by this module.
