+++ cmd/snap-update-ns/system.go
@@ -75,41 +75,6 @@ func (ctx *SystemProfileUpdateContext) Assumptions() *Assumptions {
 	return as
 }
 
-func applySystemFstab(ctx MountProfileUpdateContext) error {
-	unlock, err := ctx.Lock()
-	if err != nil {
-		return err
-	}
-	defer unlock()
-
-	// Read the desired and current mount profiles. Note that missing files
-	// count as empty profiles so that we can gracefully handle a mount
-	// interface connection/disconnection.
-	desired, err := ctx.LoadDesiredProfile()
-	if err != nil {
-		return err
-	}
-	debugShowProfile(desired, "desired mount profile")
-
-	currentBefore, err := ctx.LoadCurrentProfile()
-	if err != nil {
-		return err
-	}
-	debugShowProfile(currentBefore, "current mount profile (before applying changes)")
-	// Synthesize mount changes that were applied before for the purpose of the tmpfs detector.
-	as := ctx.Assumptions()
-	for _, entry := range currentBefore.Entries {
-		as.AddChange(&Change{Action: Mount, Entry: entry})
-	}
-
-	currentAfter, err := applyProfile(ctx, currentBefore, desired, as)
-	if err != nil {
-		return err
-	}
-
-	return ctx.SaveCurrentProfile(currentAfter)
-}
-
 // desiredSystemProfilePath returns the path of the fstab-like file with the desired, system-wide mount profile for a snap.
 func desiredSystemProfilePath(snapName string) string {
 	return fmt.Sprintf("%s/snap.%s.fstab", dirs.SnapMountPolicyDir, snapName)
