+++ cmd/snap-update-ns/utils.go
@@ -653,50 +653,3 @@ func createWritableMimic(dir, neededBy string, as *Assumptions) ([]*Change, erro
 	}
 	return changes, nil
 }
-
-func applyProfile(up MountProfileUpdateContext, currentBefore, desired *osutil.MountProfile, as *Assumptions) (*osutil.MountProfile, error) {
-	// Compute the needed changes and perform each change if
-	// needed, collecting those that we managed to perform or that
-	// were performed already.
-	changesNeeded := NeededChanges(currentBefore, desired)
-	debugShowChanges(changesNeeded, "mount changes needed")
-
-	logger.Debugf("performing mount changes:")
-	var changesMade []*Change
-	for _, change := range changesNeeded {
-		logger.Debugf("\t * %s", change)
-		synthesised, err := change.Perform(as)
-		changesMade = append(changesMade, synthesised...)
-		if len(synthesised) > 0 {
-			logger.Debugf("\tsynthesised additional mount changes:")
-			for _, synth := range synthesised {
-				logger.Debugf(" * \t\t%s", synth)
-			}
-		}
-		if err != nil {
-			// We may have done something even if Perform itself has
-			// failed. We need to collect synthesized changes and
-			// store them.
-			origin := change.Entry.XSnapdOrigin()
-			if origin == "layout" || origin == "overname" {
-				return nil, err
-			} else if err != ErrIgnoredMissingMount {
-				logger.Noticef("cannot change mount namespace according to change %s: %s", change, err)
-			}
-			continue
-		}
-
-		changesMade = append(changesMade, change)
-	}
-
-	// Compute the new current profile so that it contains only changes that were made
-	// and save it back for next runs.
-	var currentAfter osutil.MountProfile
-	for _, change := range changesMade {
-		if change.Action == Mount || change.Action == Keep {
-			currentAfter.Entries = append(currentAfter.Entries, change.Entry)
-		}
-	}
-	debugShowProfile(&currentAfter, "current mount profile (after applying changes)")
-	return &currentAfter, nil
-}
